nero-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
python-jose
apscheduler
pydantic
openai
